staging: Rename content buffer module
Target content construction no longer shares a module with index mutation.
The staging.operations basename now reads like a catch-all after the
index-update split.

Keeping the generic name would make staging look broader than it is.
Callers only need builders for index and worktree target buffers.

This commit moves those builders to staging.content_buffers and rewrites
command imports to name the narrower module directly.

The staging package now presents separate modules for content construction
and Git index mutation.
